+++ ordiniBarScuolaBorsa/__pycache__/api.cpython-312.pyc
