+++ ordiniBarScuolaBorsa/index.py
@@ -1,0 +1,5 @@
+from ordiniBarScuolaBorsa import app,render_template
+
+@app.route('/')
+def index():
+    return render_template('index.html')
