+++ ordiniBarScuolaBorsa/menu.py
@@ -1,0 +1,5 @@
+from ordiniBarScuolaBorsa import app,render_template
+
+@app.route('/menu')
+def orders():
+    return render_template('menu.html')
