+++ ordiniBarScuolaBorsa/old_app/__pycache__/__init__.cpython-312.pyc
