+++ ordiniBarScuolaBorsa/old_app/utils.py
