+++ ordiniBarScuolaBorsa/orders.py
@@ -1,0 +1,5 @@
+from ordiniBarScuolaBorsa import app,render_template
+
+@app.route('/orders')
+def orders():
+    return render_template('orders.html')
