+++ ordiniBarScuolaBorsa/queue copy.py
@@ -1,0 +1,5 @@
+from ordiniBarScuolaBorsa import app,render_template
+
+@app.route('/admin')
+def index():
+    return render_template('admin.html')
