+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "ordiniBarScuolaBorsa"
+version = "1.0.0"
+description = "ordini del bar della scuola borsa"
+dependencies = [
+    "flask"
+]
+
+[build-system]
+requires = ["flit_core<5"]
+build-backend = "flit_core.buildapi"
